EC-796 Refactor components and add tests
