--- Comments/ClientApp/src/components/Questions/fake-template-data.json
@@ -1,30 +1,0 @@
-[
-  {
-    "_id": "5d1df8d7c9a51d6d8aacfac1",
-    "question": "Do you agree with the proposal not to update the guideline?"
-  },
-  {
-    "_id": "5d1df8d7ba80fa6498b5cc7c",
-    "question": "Do you have any comments on areas excluded from the scope of the guideline?"
-  },
-  {
-    "_id": "5d1df8d733eb5122a97d46ef",
-    "question": "Do you have any comments on equalities issues?"
-  },
-  {
-    "_id": "5d1df8d786b906d5e94788a6",
-    "question": "Has all of the relevant evidence been taken into account?"
-  },
-  {
-    "_id": "5d1df8d7e898e609e1b92834",
-    "question": "Are the summaries of clinical and cost effectiveness reasonable interpretations of the evidence?"
-  },
-  {
-    "_id": "5d1df8d7e898e609e1b928343",
-    "question": "Are the recommendations sound and a suitable basis for guidance to the NHS?"
-  },
-  {
-    "_id": "5d1df8d7e898e609e1b928344",
-    "question": "Are there any aspects of the recommendations that need particular consideration to ensure we avoid unlawful discrimination against any group of people on the grounds of race, gender, disability, religion or belief, sexual orientation, age, gender reassignment, pregnancy and maternity?"
-  }
-]
