--- Comments/ClientApp/src/components/Questions/fake-template-data2.json
@@ -1,74 +1,0 @@
-{
-  "questions": [
-    {
-        "questionId": 4,
-        "questionText": "Which areas will have the biggest impact on practice and be challenging to implement? Please say for whom and why.",
-        "questionTypeId": 1,
-        "questionOrder": 1,
-        "lastModifiedByUserId": "00000000-0000-0000-0000-000000000000",
-        "lastModifiedDate": "2018-08-03T22:44:58.45",
-        "questionType": {
-            "description": "A text question requiring a text answer.",
-            "hasTextAnswer": true,
-            "hasBooleanAnswer": false,
-            "type":"Text"
-        },
-        "locationId": 7,
-        "htmlElementID": null,
-        "rangeStart": null,
-        "rangeStartOffset": null,
-        "rangeEnd": null,
-        "rangeEndOffset": null,
-        "quote": null,
-        "commentOn": "Consultation",
-        "documentId": null,
-        "show": true
-    },
-    {
-        "questionId": 5,
-        "questionText": "Would implementation of any of the draft recommendations have significant cost implications?",
-        "questionTypeId": 1,
-        "questionOrder": 2,
-        "lastModifiedByUserId": "00000000-0000-0000-0000-000000000000",
-        "lastModifiedDate": "2018-08-03T22:44:58.45",
-        "questionType": {
-            "description": "A text question requiring a text answer.",
-            "hasTextAnswer": true,
-            "hasBooleanAnswer": false,
-            "type":"Text"
-        },
-        "locationId": 8,
-        "htmlElementID": null,
-        "rangeStart": null,
-        "rangeStartOffset": null,
-        "rangeEnd": null,
-        "rangeEndOffset": null,
-        "quote": null,
-        "commentOn": "Consultation",
-        "documentId": 1
-    },
-    {
-        "questionId": 6,
-        "questionText": "Do you like rainbows?",
-        "questionTypeId": 1,
-        "questionOrder": 3,
-        "lastModifiedByUserId": "00000000-0000-0000-0000-000000000000",
-        "lastModifiedDate": "2018-08-03T22:44:58.45",
-        "questionType": {
-            "description": "A yes or no question.",
-            "hasTextAnswer": true,
-            "hasBooleanAnswer": true,
-            "type":"YesNo"
-        },
-        "locationId": 9,
-        "htmlElementID": null,
-        "rangeStart": null,
-        "rangeStartOffset": null,
-        "rangeEnd": null,
-        "rangeEndOffset": null,
-        "quote": null,
-        "commentOn": "Consultation",
-        "documentId": 2
-    }
-  ]
-}
